+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## HEAD (Unreleased)
 * Upgrade to terraform-bridge 3.11.0
 * Upgrade to pulumi 3.17.0
+* Avoid blocking the message loop for Docker builds in Python
+  [#314](https://github.com/pulumi/pulumi-docker/pull/314).
 
 ---